PP-9955 refactor error handling  structure
- refactor error-handling structure to follow GDS' project directory structure
