+++ app/error-handling/404.njk
@@ -1,4 +1,4 @@
-{% extends "./layout.njk" %}
+{% extends "../views/layout.njk" %}
 
 {% block pageTitle %}
   {{ __p('404Page.title') }} - GOV.UK Pay
