+++ app/error-handling/500.njk
@@ -1,4 +1,4 @@
-{% extends "./layout.njk" %}
+{% extends "../views/layout.njk" %}
 
 {% block pageTitle %}
   {{ __p('error.title') }} - GOV.UK Pay
